refactor: centralize gi module wrappers
