+++ ulauncher/app.py
@@ -5,11 +5,10 @@ import logging
 from typing import Any, Iterable, Literal, cast
 from weakref import WeakValueDictionary
 
-from gi.repository import Gdk, Gio, Gtk
-
 import ulauncher
 from ulauncher import app_id, first_run
 from ulauncher.cli import get_cli_args
+from ulauncher.gi import Gdk, Gio, Gtk
 from ulauncher.internals.result import Result
 from ulauncher.ui.windows.preferences.preferences_window import PreferencesWindow
 from ulauncher.ui.windows.ulauncher_window import UlauncherWindow
