+++ ulauncher/ui/windows/preferences/preferences_window.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from typing import Any
 
-from gi.repository import Gdk, Gtk
-
 from ulauncher import paths
+from ulauncher.gi import Gdk, Gtk
 from ulauncher.ui.windows.preferences.views import BaseView, styled
 from ulauncher.ui.windows.preferences.views.about import AboutView
 from ulauncher.ui.windows.preferences.views.extensions import ExtensionsView
