+++ ulauncher/ui/windows/preferences/utils/ext_handlers.py
@@ -5,9 +5,8 @@ import logging
 import threading
 from typing import Callable
 
-from gi.repository import GLib, Gtk
-
 from ulauncher import paths
+from ulauncher.gi import GLib, Gtk
 from ulauncher.modes.extensions import ext_exceptions
 from ulauncher.modes.extensions.extension_controller import ExtensionController
 from ulauncher.ui.windows.preferences.views import DialogLauncher, get_window_for_widget, styled
