+++ ulauncher/ui/windows/preferences/views/about.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-from gi.repository import Gtk
-
 from ulauncher import api_version, paths, version
+from ulauncher.gi import Gtk
 from ulauncher.ui.windows.preferences.views import BaseView, styled
 from ulauncher.utils.launch_detached import open_detached
 from ulauncher.utils.load_icon_surface import load_icon_surface
