+++ ulauncher/ui/windows/ulauncher_window.py
@@ -3,10 +3,9 @@ from __future__ import annotations
 import logging
 from typing import Any, Iterable
 
-from gi.repository import Gdk, GLib, Gtk
-
 from ulauncher import paths
 from ulauncher.core import UlauncherCore
+from ulauncher.gi import Gdk, GLib, Gtk
 from ulauncher.internals.result import Result
 from ulauncher.ui import layer_shell
 from ulauncher.ui.item_navigation import ItemNavigation
