+++ ulauncher/utils/hotkey_controller.py
@@ -4,9 +4,8 @@ import logging
 import subprocess
 from shutil import which
 
-from gi.repository import Gio, GLib
-
 from ulauncher import app_id
+from ulauncher.gi import Gio, GLib
 from ulauncher.ui.windows.hotkey_dialog import HotkeyDialog
 from ulauncher.utils.environment import DESKTOP_ID, DESKTOP_NAME
 from ulauncher.utils.launch_detached import launch_detached
